Updated Prisma Schema relations referential actions

+++ app/api/stores/[storeId]/orders/[orderId]/route.ts
@@ -2,15 +2,13 @@ import prisma from "@/prisma/client";
 import { auth } from "@clerk/nextjs/server";
 import { NextResponse } from "next/server";
 
-
-
 export async function GET(req: Request, { params }: { params: { storeId: string; orderId: string } }) {
   try {
-    // const { userId } = auth();
-    
-    // if (!userId) {
-    //   return new NextResponse("User not found", { status: 403 });
-    // }
+    const { userId } = auth();
+
+    if (!userId) {
+      return new NextResponse("User not found", { status: 403 });
+    }
     if (!params.orderId) {
       return new NextResponse("Order id is required", { status: 400 });
     }
@@ -18,13 +16,13 @@ export async function GET(req: Request, { params }: { params: { storeId: string;
       return new NextResponse("Store id is required", { status: 400 });
     }
 
-    // const storeByUserId = await prisma.store.findFirst({
-    //   where: { id: params.storeId, userId },
-    // });
+    const storeByUserId = await prisma.store.findFirst({
+      where: { id: params.storeId, userId },
+    });
 
-    // if (!storeByUserId) {
-    //   return new NextResponse("Unauthorized", { status: 403 });
-    // }
+    if (!storeByUserId) {
+      return new NextResponse("Unauthorized", { status: 403 });
+    }
 
     const order = await prisma.order.findUnique({
       where: { id: params.orderId },
@@ -32,8 +30,7 @@ export async function GET(req: Request, { params }: { params: { storeId: string;
         orderItems: { include: { product: true, productVariation: true } },
       },
     });
-    console.log("order: ", order);
-    return NextResponse.json(order, );
+    return NextResponse.json(order);
   } catch (error) {
     console.trace("[ORDER_GET]", error);
     return new NextResponse("Internal error", { status: 500 });
@@ -42,42 +39,37 @@ export async function GET(req: Request, { params }: { params: { storeId: string;
 
 export async function PATCH(req: Request, { params }: { params: { storeId: string; orderId: string } }) {
   try {
-    // const { userId } = auth();
+    const { userId } = auth();
     const body = await req.json();
     const orderData = body;
-    console.log('orderData: ', orderData);
 
-    // if (!userId) {
-    //   return new NextResponse("User not found", { status: 403 });
-    // }
+    if (!userId) {
+      return new NextResponse("User not found", { status: 403 });
+    }
     if (!params.orderId) {
       return new NextResponse("Order id is required", { status: 400 });
     }
     if (!params.storeId) {
       return new NextResponse("Store id is required", { status: 400 });
     }
-    if (!orderData.orderItems) {
-      return new NextResponse("Order Items are required", { status: 400 });
-    }
 
     if (!orderData.phone || !orderData.address) {
       return new NextResponse("Order details are required", { status: 400 });
     }
 
-    // const storeByUserId = await prisma.store.findFirst({
-    //   where: { id: params.storeId, userId },
-    // });
+    const storeByUserId = await prisma.store.findFirst({
+      where: { id: params.storeId, userId },
+    });
 
-    // if (!storeByUserId) {
-    //   return new NextResponse("Unauthorized", { status: 403 });
-    // }
+    if (!storeByUserId) {
+      return new NextResponse("Unauthorized", { status: 403 });
+    }
 
     const updatedOrder = await prisma.order.update({
       where: { id: params.orderId },
       data: orderData,
     });
 
-    console.log('updatedOrder: ', updatedOrder);
     return NextResponse.json(updatedOrder);
   } catch (error) {
     console.trace("[ORDER_PATCH]", error);
@@ -87,22 +79,22 @@ export async function PATCH(req: Request, { params }: { params: { storeId: strin
 
 export async function DELETE(req: Request, { params }: { params: { storeId: string; orderId: string } }) {
   try {
-    // const { userId } = auth();
-    // if (!userId) {
-    //   return new NextResponse("Unauthorized", { status: 404 });
-    // }
+    const { userId } = auth();
+    if (!userId) {
+      return new NextResponse("Unauthorized", { status: 404 });
+    }
     if (!params.orderId) {
       return new NextResponse("Order id is required", { status: 400 });
     }
     if (!params.storeId) {
       return new NextResponse("Store id is required", { status: 400 });
     }
-    // const storeByUserId = await prisma.store.findFirst({
-    //   where: { id: params.storeId, userId },
-    // });
-    // if (!storeByUserId) {
-    //   return new NextResponse("Unauthorized", { status: 403 });
-    // }
+    const storeByUserId = await prisma.store.findFirst({
+      where: { id: params.storeId, userId },
+    });
+    if (!storeByUserId) {
+      return new NextResponse("Unauthorized", { status: 403 });
+    }
     const deletedOrder = await prisma.order.delete({
       where: { id: params.orderId },
     });
